refactor(domain): remove unnecessary type annotations in Comment entities
Remove explicit type annotations where TypeScript can infer them
